시큐리티 설정 및 로그인

+++ src/main/java/com/example/sns_project/entity/User.java
@@ -9,7 +9,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class User {
     
-    private int userNum;
+    private Long userNum;
     
     private String id;
